Added code to find min. and max. in array
